// src/app/Models/journal-report.model.ts
import { IdDescription } from '../interfaces/id-description';

export class JournalReportInputHelp {
  GLcodeList: IdDescription[];
  TrnTypeList: IdDescription[];
  TrnModeList: IdDescription[];
  TrnModeListDailyTransactionReport: IdDescription[];
  TransactionTypeList: IdDescription[];
  NormalTransactionTypeList: IdDescription[];
  TrnNatureListDailyTransactionReport: IdDescription[];
}